feat(user): improve test
- modify test name pattern
  - HttpMethod URL - response
- setup fixtures by using faker

+++ test/api/users.spec.ts
@@ -1,27 +1,70 @@
 import request from 'supertest'
-import { createConnection, getConnection } from 'typeorm'
+import faker from 'faker'
+import { createConnection, Connection } from 'typeorm'
 import { app } from '~/server/app'
 import { User } from '~/server/entity'
 import { UserUpdateInfo } from '~/server/services'
 
 describe('API - users', () => {
   const req = request(app)
+  let conn: Connection
+  let user: User
 
-  beforeAll(() => createConnection())
+  const registInfo = {
+    email: faker.internet.email(),
+    username: faker.internet.userName(),
+    password: faker.internet.password()
+  }
 
-  afterAll(() =>
-    getConnection()
-      .createQueryBuilder()
-      .delete()
-      .from(User)
-      .execute())
+  const loginInfo = {
+    email: registInfo.email,
+    password: registInfo.password
+  }
 
-  test('should return user for regist', async () => {
+  beforeAll(async () => {
+    conn = await createConnection()
+    if (!conn || !conn.isConnected) {
+      fail('cannot connect database')
+    }
+  })
+
+  beforeEach(async () => {
+    if (!conn || !conn.isConnected) {
+      fail('cannot connect database')
+    }
+
+    user = await conn.getRepository(User)
+      .save(registInfo)
+  })
+
+  afterAll(async () => {
+    if (conn && conn.isConnected) {
+      await conn.createQueryBuilder()
+        .delete()
+        .from(User)
+        .execute()
+
+      await conn.close()
+    }
+  })
+
+  async function getAuthentication(): Promise<string> {
+    const res = await req.post('/api/users/login')
+      .send({ user: loginInfo })
+      .expect(200)
+
+    expect(res.body).toHaveProperty('user')
+    expect(res.body.user.token).not.toBeNull()
+
+    return res.body.user.token
+  }
+
+  test('post /api/users - returns User', async () => {
     // Given
     const user = {
-      email: 'login@test.com',
-      username: 'username',
-      password: 'Secret!'
+      email: faker.internet.email(),
+      username: faker.internet.userName(),
+      password: faker.internet.password()
     }
 
     // When
@@ -36,30 +79,60 @@ describe('API - users', () => {
     expect(res.body.user.token).not.toBeNull()
   })
 
-  test('should return user for login', async () => {
+  test('post /api/users - throws UnprocessableEntityError #1', async () => {
     // Given
     const user = {
-      email: 'login@test.com',
-      password: 'Secret!'
+      email: faker.internet.email(),
+      password: faker.internet.password()
     }
 
     // When
-    const res = await req.post('/api/users/login')
+    const res = await req.post('/api/users')
       .send({ user })
+      .expect(422)
+
+    // Then
+    expect(res.body).toHaveProperty('errors')
+    expect(res.body.errors).toHaveProperty('body')
+    expect(res.body.errors.body).toEqual(['Name is required'])
+  })
+
+  test('post /api/users - throws UnprocessableEntityError #2', async () => {
+      // Given
+    const user = {
+      username: faker.internet.userName(),
+      password: faker.internet.password()
+    }
+
+      // When
+    const res = await req.post('/api/users')
+        .send({ user })
+        .expect(422)
+
+      // Then
+    expect(res.body).toHaveProperty('errors')
+    expect(res.body.errors).toHaveProperty('body')
+    expect(res.body.errors.body).toEqual(['Email is required'])
+  })
+
+  test('post /api/users/login - returns User', async () => {
+    // When
+    const res = await req.post('/api/users/login')
+      .send({ user: loginInfo })
       .expect(200)
 
     // Then
     expect(res.body).toHaveProperty('user')
     expect(res.body.user).toHaveProperty('email')
-    expect(res.body.user.email).toBe(user.email)
+    expect(res.body.user.email).toBe(loginInfo.email)
     expect(res.body.user.token).not.toBeNull()
   })
 
-  test('should return UnprocessableEntityError', async () => {
+  test('post /api/users/login - throws UnprocessableEntityError #1', async () => {
     // Given
     const user = {
-      email: 'login@test.com',
-      password: 'invalid password'
+      email: faker.internet.email(),
+      password: loginInfo.password
     }
 
     // When
@@ -73,23 +146,30 @@ describe('API - users', () => {
     expect(res.body.errors.body).toEqual(['email or password is invalid'])
   })
 
-  test('should return user w/ authorization', async () => {
+  test('post /api/users/login - throws UnprocessableEntityError #2', async () => {
     // Given
     const user = {
-      email: 'login@test.com',
-      password: 'Secret!'
+      email: loginInfo.email,
+      password: faker.internet.password()
     }
 
     // When
-    let res = await req.post('/api/users/login')
+    const res = await req.post('/api/users/login')
       .send({ user })
-      .expect(200)
+      .expect(422)
 
-    expect(res.body).toHaveProperty('user')
-    expect(res.body.user.token).not.toBeNull()
+    // Then
+    expect(res.body).toHaveProperty('errors')
+    expect(res.body.errors).toHaveProperty('body')
+    expect(res.body.errors.body).toEqual(['email or password is invalid'])
+  })
 
-    res = await req.get('/api/user')
-      .set('Authorization', `Token ${res.body.user.token}`)
+  test('get /api/user - returns User w/ token', async () => {
+    // When
+    const token = await getAuthentication()
+
+    const res = await req.get('/api/user')
+      .set('Authorization', `Token ${token}`)
       .expect(200)
 
     // Then
@@ -99,7 +179,7 @@ describe('API - users', () => {
     expect(res.body.user.token).not.toBeNull()
   })
 
-  test('should throw AuthorizedRequiredError #1', async () => {
+  test('get /api/user - throws Unauthorized', async () => {
     // When
     const res = await req.get('/api/user')
       .expect(401)
@@ -112,7 +192,7 @@ describe('API - users', () => {
       .toContain('Authorization is required for request on GET /api/user')
   })
 
-  test('should throw AuthorizedRequiredError #2', async () => {
+  test('put /api/user - throws Unauthorized', async () => {
     // When
     const res = await req.put('/api/user')
       .expect(401)
@@ -125,31 +205,19 @@ describe('API - users', () => {
       .toContain('Authorization is required for request on PUT /api/user')
   })
 
-  test('update user and returns loggedUser', async () => {
+  test('put /api/user - return User', async () => {
     // Given
-    const user = {
-      email: 'login@test.com',
-      password: 'Secret!'
-    }
-
-    let res = await req.post('/api/users/login')
-      .send({ user })
-      .expect(200)
-
-    expect(res.body).toHaveProperty('user')
-    expect(res.body.user.token).not.toBeNull()
-
-    const loggedUser = res.body.user
+    const token = await getAuthentication()
 
     const updateUser: UserUpdateInfo = {
-      username: `${loggedUser.username}!`,
-      email: `${loggedUser.email}!`,
+      username: `${registInfo.username}!`,
+      email: `${registInfo.email}!`,
     }
 
     // When
-    res = await req.put('/api/user')
+    const res = await req.put('/api/user')
       .send({ user: updateUser })
-      .set('Authorization', `Token ${loggedUser.token}`)
+      .set('Authorization', `Token ${token}`)
       .expect(200)
 
     // Then
@@ -158,7 +226,7 @@ describe('API - users', () => {
     expect(res.body.user.username).toEqual(updateUser.username)
     expect(res.body.user.email).toEqual(updateUser.email)
     expect(res.body.user.token).not.toBeNull()
-    expect(res.body.user.token).not.toEqual(loggedUser.token)
+    expect(res.body.user.token).not.toEqual(token)
 
   })
 })
